refactor: ディレクトリ構造を変更
