+++ frontend/src/components/auth/SignUp.tsx
@@ -1,12 +1,12 @@
 import { useContext, useState } from "react";
-import { signIn } from "../api/Auth";
+import { signIn } from "../../api/Auth";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
-import { createUser } from "../api/User";
+import { createUser } from "../../api/User";
 import { toast } from "react-toastify";
-import { extractErrorMessage } from "../utils/extractErrorMessage";
-import { UserContext } from "../contexts/UserContext";
-import { apiClient } from "../hooks/useAxiosIntercepter";
+import { extractErrorMessage } from "../../utils/extractErrorMessage";
+import { UserContext } from "../../contexts/UserContext";
+import { apiClient } from "../../hooks/useAxiosIntercepter";
 
 const SignUp = () => {
   const navigate = useNavigate();
