+++ frontend/src/components/contents/Contents.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PostList from "./PostList";
+import PostList from "./parts/PostList";
 
 const Contents = () => {
   return <PostList></PostList>;
